+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.9.0] - Apr, 2025
+- Changed the methods `Process` and `ProcessAsync` to `Match` and `MatchAsync` to adhere to the functional programming
+  paradigm. Sorry about that. 😂 The `Process` and `ProcessAsync` methods are still available, but they are marked as 
+  obsolete and will be removed in the future;
+
 ## [1.8.0] - Sep, 2024
 - Added `WithMessage` to the `Error` class to allow the error message to be changed. It will allow for more flexibility when
   handling errors;